- Removed un-needed bwtoolkitframework from the Olearia frameworks folder. - Removed un-needed talkingbooktypes from all projects - Finally got logic correct for auto play of text only files (2002 formats only checked so far) - Fixed a couple of possible conversion issues in the NSstring categories - removed a whole lot of duplicated properties from several subclasses as they were passed through from their superclass anyway. - fixed several possible ambiguous logic errors - fixed a few memory leaks. - tested and fixed issues compiling under snow leopard - changed all projects except the plugins to compile 32/64 bit release variants.  This is due to the lack of 64bit calls in the QTmovie class under leopard 10.5. This issue does not exist under Snow Leopard 10.6. - Various other improvements.

+++ trunk/Plugins/Source/NSString-TBAdditions.m
@@ -86,18 +86,18 @@
 		{
 			if(capt4 < 1000) // milliseconds value less than 1 second?
 			{
-				totalTimeInSeconds = (int)(capt4 * 10);
-				fractions = (int)(capt5 * 10); // leftover milliseconds 
+				totalTimeInSeconds = (NSInteger)(capt4 * 10);
+				fractions = (NSInteger)(capt5 * 10); // leftover milliseconds 
 			}
 			else 
 			{	// more than 1 second of milliseconds 
-				totalTimeInSeconds = (int)(capt4 / 1000);
-				fractions = (int)(capt4 % 1000) + (int)(capt5 * 10); // leftover milliseconds 
+				totalTimeInSeconds = (NSInteger)(capt4 / 1000);
+				fractions = (NSInteger)(capt4 % 1000) + (int)(capt5 * 10); // leftover milliseconds 
 			}
 		}
-		hours = totalTimeInSeconds / 3600; 
-		minutes = (totalTimeInSeconds / 3600) % 60; 
-		seconds = ((totalTimeInSeconds / 3600) % 60) % 60;
+		hours = (NSInteger)totalTimeInSeconds / 3600; 
+		minutes = (NSInteger)(totalTimeInSeconds / 3600) % 60; 
+		seconds = (NSInteger)((totalTimeInSeconds / 3600) % 60) % 60;
 	}
 	else 
 	{
